Update build.gradle
Updates the task to the more modern tasks.register, registers it under the publishing group, and adds a description. Also moved curl to default to Windows curl for Windows OSes, as this is always part of the Windows OS. Other OSes should be fine with curl but I don't have a machine to test them with.